memes files

+++ algebre.h
@@ -9,19 +9,19 @@
 
 
 
-float* produit_MV( float* vecteur) ;
+double* produit_MV( double* vecteur) ;
 
-float produitScalaire(float *A, float *B, int taille);
+double produitScalaire(double *A, double *B, int taille);
 
-void copierTableau(float *source, float *destination, int taille); 
+void copierTableau(double *source, double *destination, int taille); 
 
-void sommeTableaux(float *A, float *B, float *C, int taille);
+void sommeTableaux(double *A, double *B, double *C, int taille);
 
-void differenceTableaux(float *A, float *B, float *C, int taille);
+void differenceTableaux(double *A, double *B, double *C, int taille);
 
-void scalaireMultiplieTableau(float scalaire, float *vecteur, float *resultat, int taille);
+void scalaireMultiplieTableau(double scalaire, double *vecteur, double *resultat, int taille);
 
-float norm(float *A, int taille) ;
+double norm(double *A, int taille) ;
 
 
 
@@ -33,7 +33,7 @@ float norm(float *A, int taille) ;
 
 int indexe(int i,int j); // focntion qui donne l'indice de l'element stocké en fonction de i et j
 int couple(int I,int axis); //focntion inverse de index qui donne i ou j selon axe choisi
-float maillage(int I,int axis); // fonction qui donne xi ou xj selon axe choisi
+double maillage(int I,int axis); // fonction qui donne xi ou xj selon axe choisi
