+++ fonction.h
@@ -2,8 +2,8 @@
 #define FONCTION_H
 
 
-float* v(float x, float y);
-float u0(float x, float y);
+double* v(double x, double y);
+double u0(double x, double y);
 
 
 
@@ -11,9 +11,9 @@ float u0(float x, float y);
 
 
 
-float alpha( float x,float y);
-float beta( float x,float y);
-float gama( float x,float y);
+double alpha( double x,double y);
+double beta( double x,double y);
+double gama( double x,double y);
